+++ packages/terra-paginator/src/ControlledProgressivePaginator.jsx
@@ -47,6 +47,12 @@ const propTypes = {
    * Allows user to set custom page label. _(usage note: User must pass translated text)_.
    */
   pageLabel: PropTypes.string,
+  /**
+   * ![IMPORTANT](https://badgen.net/badge/UX/Accessibility/blue)
+   * When specified allows user to set custom page count. User should provide custom page count as part `pageLabel` for best accessibility practices.
+   * _(usage note: when `pageLabel` is not provided page count will not be hidden and default page count is displayed for best accessibility practices)_.
+   */
+  hidePageCount: PropTypes.bool,
 };
 
 class ControlledProgressivePaginator extends React.Component {
@@ -105,6 +111,7 @@ class ControlledProgressivePaginator extends React.Component {
       ariaLabelledBy,
       ariaLabel,
       pageLabel,
+      hidePageCount,
     } = this.props;
     const totalPages = (totalCount) ? calculatePages(totalCount, itemCountPerPage) : 0;
     const previousPageIndex = selectedPage === 1 ? 1 : selectedPage - 1;
@@ -113,14 +120,11 @@ class ControlledProgressivePaginator extends React.Component {
     const renderFirstandLastButton = totalCount > 0;
 
     const { messageId, messageAttributes } = getPageLabel(pageLabel, selectedPage, totalPages);
+    const pageDetails = (hidePageCount && pageLabel) ? <div>{pageLabel}</div> : <div>{intl.formatMessage({ id: messageId }, messageAttributes)}</div>;
 
     const fullViewChildren = (
       <>
-        <div>
-          {
-            intl.formatMessage({ id: messageId }, messageAttributes)
-          }
-        </div>
+        {pageDetails}
         <ul className={cx('progressive-list')}>
           {
             renderFirstandLastButton && (
@@ -212,6 +216,7 @@ class ControlledProgressivePaginator extends React.Component {
       ariaLabelledBy,
       ariaLabel,
       pageLabel,
+      hidePageCount,
     } = this.props;
     const totalPages = (totalCount) ? calculatePages(totalCount, itemCountPerPage) : 0;
     const previousPageIndex = selectedPage === 1 ? 1 : selectedPage - 1;
@@ -220,14 +225,11 @@ class ControlledProgressivePaginator extends React.Component {
     const renderFirstandLastButton = totalCount > 0;
 
     const { messageId, messageAttributes } = getPageLabel(pageLabel, selectedPage, totalPages);
+    const pageDetails = (hidePageCount && pageLabel) ? <div>{pageLabel}</div> : <div>{intl.formatMessage({ id: messageId }, messageAttributes)}</div>;
 
     const reducedViewChildren = (
       <>
-        <div>
-          {
-            intl.formatMessage({ id: messageId }, messageAttributes)
-          }
-        </div>
+        {pageDetails}
         <ul className={cx('progressive-list')}>
           {
             renderFirstandLastButton && (