+++ packages/terra-paginator/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added `hidePageCount` prop to hide page count provided by paginator.
+
 ## 2.85.0 - (September 19, 2023)
 
 * Added